Move `checkInstalled` to method on runtime
