+++ extensions/positron-r/src/runtime.ts
@@ -8,10 +8,10 @@ import PQueue from 'p-queue';
 
 import { JupyterAdapterApi, JupyterKernelSpec, JupyterLanguageRuntime, JupyterKernelExtra } from './jupyter-adapter';
 import { ArkLsp, LspState } from './lsp';
-import { delay } from './util';
+import { delay, timeout } from './util';
 import { ArkAttachOnStartup, ArkDelayStartup } from './startup';
 import { RHtmlWidget, getResourceRoots } from './htmlwidgets';
-import { getRPackageName } from './contexts';
+import { randomUUID } from 'crypto';
 
 export let lastRuntimePath = '';
 
@@ -289,6 +289,71 @@ export class RRuntime implements positron.LanguageRuntime, vscode.Disposable {
 	 */
 	showOutput() {
 		this._kernel?.showOutput();
+	}
+
+	/**
+	 * Checks whether a package is installed in the runtime.
+	 * @param pkgName The name of the package to check
+	 * @param intendedUse A string describing the intended use of the package, for a warning message
+	 * @returns true if the package is installed, false otherwise
+	 */
+
+	async checkInstalled(pkgName: string, intendedUse: string, pkgVersion?: string): Promise<boolean> {
+		let isInstalled: boolean;
+		try {
+			if (pkgVersion) {
+				isInstalled = await this.callMethod('is_installed', pkgName, pkgVersion);
+			} else {
+				isInstalled = await this.callMethod('is_installed', pkgName);
+			}
+		} catch (err) {
+			const runtimeError = err as positron.RuntimeMethodError;
+			throw new Error(`Error checking for package ${pkgName}: ${runtimeError.message} ` +
+				`(${runtimeError.code})`);
+		}
+
+		if (!isInstalled) {
+			const install = await positron.window.showSimpleModalDialogPrompt(
+				'Missing R package',
+				`Package ${pkgName} required but not installed.`,
+				'Install now'
+			);
+			if (install) {
+				const id = randomUUID();
+
+				// A promise that resolves when the runtime is idle:
+				const promise = new Promise<void>(resolve => {
+					const disp = this.onDidReceiveRuntimeMessage(runtimeMessage => {
+						if (runtimeMessage.parent_id === id &&
+							runtimeMessage.type === positron.LanguageRuntimeMessageType.State) {
+							const runtimeMessageState = runtimeMessage as positron.LanguageRuntimeState;
+							if (runtimeMessageState.state === positron.RuntimeOnlineState.Idle) {
+								resolve();
+								disp.dispose();
+							}
+						}
+					});
+				});
+
+				this.execute(`install.packages("${pkgName}")`,
+					id,
+					positron.RuntimeCodeExecutionMode.Interactive,
+					positron.RuntimeErrorBehavior.Continue);
+
+				// Wait for the the runtime to be idle, or for the timeout:
+				await Promise.race([promise, timeout(2e4, 'waiting for package installation')]);
+
+				return true;
+			} else {
+				if (pkgVersion) {
+					vscode.window.showWarningMessage(`Cannot ${intendedUse} without ${pkgName} ${pkgVersion}.`);
+				} else {
+					vscode.window.showWarningMessage(`Cannot ${intendedUse} without ${pkgName} package.`);
+				}
+				return false;
+			}
+		}
+		return true;
 	}
 
 	private async createKernel(): Promise<JupyterLanguageRuntime> {
